fixed error in MyBookRepository

+++ src/main/java/com/bookstore/GeekText/repository/MyBookRepository.java
@@ -4,6 +4,6 @@ import com.bookstore.GeekText.model.Books;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface MyBookRepository extends JpaRepository<Books, Integer> {
+public interface MyBookRepository extends JpaRepository<Books, Long> {
 
 }
